getting close. Need to move the cached track out of the api, and have it part of the trait interface
